to move generator.go and limiter.go to their own directories under lib
