+++ cmd/gentx.go
@@ -19,8 +19,9 @@ var gentxCmd = &cobra.Command{
 		senderCount, _ := cmd.Flags().GetInt("sender-count")
 		txCount, _ := cmd.Flags().GetInt("tx-count")
 		txStoreDir, _ := cmd.Flags().GetString("tx-store-dir")
+		mempool, _ := cmd.Flags().GetInt("mempool")
 
-		gentx.GenTx(httpRpc, faucetPrivateKey, senderCount, txCount, txStoreDir)
+		gentx.GenTx(httpRpc, faucetPrivateKey, senderCount, txCount, txStoreDir, mempool)
 		fmt.Println("gentx called")
 	},
 }
