+++ lib/cmd/gentx/gentx.go
@@ -3,11 +3,14 @@ package gentx
 import (
 	"log"
 
-	"github.com/0glabs/evmchainbench/lib/cmd/run"
+	generatorpkg "github.com/0glabs/evmchainbench/lib/generator"
+	limiterpkg "github.com/0glabs/evmchainbench/lib/limiter"
 )
 
-func GenTx(rpcUrl, faucetPrivateKey string, senderCount, txCount int, txStoreDir string) {
-	generator, err := run.NewGenerator(rpcUrl, faucetPrivateKey, senderCount, txCount, true, txStoreDir, nil)
+func GenTx(rpcUrl, faucetPrivateKey string, senderCount, txCount int, txStoreDir string, mempool int) {
+	limiter := limiterpkg.NewRateLimiter(mempool)
+
+	generator, err := generatorpkg.NewGenerator(rpcUrl, faucetPrivateKey, senderCount, txCount, true, txStoreDir, limiter)
 	if err != nil {
 		log.Fatalf("Failed to create generator: %v", err)
 	}
